requirements.txt
dbt-duckdb
pqdm
pandas
json_tricks
openpyxl
streamlit
watchdog
